Move index.js code into modules
